pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-metaminds-er-triage"
version = "0.1.0"
description = "METAMINDS ER triage environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "click>=8.1.8,<9.0",
    "fastapi>=0.115.2,<1.0",
    "uvicorn[standard]>=0.35.0,<1.0",
    "pydantic>=2.11.7,<3.0",
    "httpx>=0.28.1,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3,<9.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
py-modules = ["client"]
packages = ["app", "server", "baselines"]
